+++ packages/thirdweb/src/wallets/in-app/core/interfaces/connector.ts
@@ -3,6 +3,7 @@ import type {
   AuthArgsType,
   AuthLoginReturnType,
   GetUser,
+  LogoutReturnType,
   PreAuthArgsType,
   SendEmailOtpReturnType,
 } from "../authentication/type.js";
@@ -12,4 +13,5 @@ export interface InAppConnector {
   getAccount(): Promise<Account>;
   preAuthenticate(args: PreAuthArgsType): Promise<SendEmailOtpReturnType>;
   authenticate(args: AuthArgsType): Promise<AuthLoginReturnType>;
+  logout(): Promise<LogoutReturnType>;
 }