+++ packages/thirdweb/src/wallets/in-app/native/helpers/auth/logout.ts
@@ -1,13 +1,14 @@
+import type { LogoutReturnType } from "../../../core/authentication/type.js";
 import {
   removeAuthTokenInClient,
   removeLoggedInWalletUserId,
 } from "../storage/local.js";
 import { getCognitoUser } from "../storage/state.js";
 
-export async function logoutUser(clientId: string): Promise<boolean> {
+export async function logoutUser(clientId: string): Promise<LogoutReturnType> {
   const isLoggedUserOutIncognito = await removeAuthTokenInClient(clientId);
   await removeLoggedInWalletUserId(clientId);
   const cognitoUser = getCognitoUser();
   cognitoUser?.signOut();
-  return isLoggedUserOutIncognito;
+  return { success: isLoggedUserOutIncognito };
 }